+++ HISTORY.rst
@@ -16,9 +16,9 @@
 - A registry can now carry metadata just like a provider.
 - A registry now has an attribute
   :attr:`skosprovider.registry.Registry.instance_scope` that indicates how the
-  registry is being run. This can help providers determine if they're
-  compatible with a registry run this was. Especially important for
-  SQLAlchemyProvider run in a web application.
+  registry is managed in the application process. This can help providers 
+  determine if they're compatible with a particular registry. Especially 
+  important for SQLAlchemyProvider run in a web application.
 - Fix a bug that made it impossible for a
   :class:`~skosprovider.providers.SimpleCsvProvider` to read sources. (#36)
 - Drop support for Python 3.3.